# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Event AI Server API
  description: |
    A comprehensive REST API for the Event AI platform that handles event management, 
    AI-powered crowd simulation, and real-time notifications.
    
    ## Architecture
    - **Server**: Express.js on AWS EC2
    - **Database**: Supabase PostgreSQL  
    - **Storage**: AWS S3
    - **AI**: Amazon SageMaker
    - **Notifications**: n8n + WhatsApp Business API
    
    ## Authentication
    Currently using API key authentication. Future versions will include JWT.
    
    ## Rate Limiting
    - DISABLED - No rate limiting applied for development
    
    ## Response Format
    All responses follow a consistent format:
    ```json
    {
      "success": true|false,
      "data": {...},
      "message": "Optional message",
      "error": "Error details if success=false"
    }
    ```
  version: 1.0.0
  contact:
    name: Event AI Team
    email: support@eventai.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://your-domain.com
    description: Production server
  - url: https://staging.your-domain.com
    description: Staging server
paths:
...
  /api/v1/events:
    post:
      summary: Create a new event
      description: Creates a new event with optional file uploads for ticketing data and seating charts
      tags:
        - Events
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - venue
                - expectedAttendees
                - eventDate
                - eventType
              properties:
                name:
                  type: string
                  description: Event name
                  example: "Summer Music Festival 2024"
                description:
                  type: string
                  description: Event description
                  example: "Annual outdoor music festival featuring top artists"
                venue:
                  type: string
                  description: Event venue
                  example: "Central Park"
                expectedAttendees:
                  type: integer
                  minimum: 1
                  description: Expected number of attendees
                  example: 5000
                eventDate:
                  type: string
                  format: date-time
                  description: Event date and time
                  example: "2024-07-15T18:00:00Z"
                eventType:
                  type: string
                  enum: [CONCERT, CONFERENCE, SPORTS, FESTIVAL, OTHER]
                  description: Type of event
                  example: "FESTIVAL"
                ticketingData:
                  type: string
                  format: binary
                  description: CSV file with ticketing data
                seatingChart:
                  type: string
                  format: binary
                  description: JSON file with seating chart data
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      eventId:
                        type: string
                        example: "evt_a03bf463-7940-4df0-a612-8e41994784e2"
                      simulationId:
                        type: string
                        example: "sim_123456789"
                  message:
                    type: string
                    example: "Event created successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      summary: Get all events
      description: Retrieves a paginated list of events
      tags:
        - Events
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of events per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: status
          in: query
          description: Filter by event status
          schema:
            type: string
            enum: [CREATED, ACTIVE, COMPLETED, CANCELLED]
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      events:
                        type: array
                        items:
                          $ref: '#/components/schemas/Event'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/events/{eventId}:
    get:
      summary: Get event by ID
      description: Retrieves a specific event by its ID
      tags:
        - Events
      parameters:
        - name: eventId
          in: path
          required: true
          description: Event ID
          schema:
            type: string
            example: "evt_a03bf463-7940-4df0-a612-8e41994784e2"
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Event'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update event
      description: |
        Updates an existing event with support for attachment URLs and context for AI chatbot knowledge base.
        All fields are optional - only provided fields will be updated.
      tags:
        - Events
      parameters:
        - name: eventId
          in: path
          required: true
          description: Event ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Event name
                  maxLength: 255
                  example: "Updated Concert in the Park"
                description:
                  type: string
                  description: Event description
                  maxLength: 1000
                  example: "An updated description of the concert event"
                venue:
                  type: string
                  description: Event venue
                  maxLength: 255
                  example: "Central Park Amphitheater"
                dateOfEventStart:
                  type: string
                  format: date-time
                  description: Event start date and time
                  example: "2024-07-15T18:00:00Z"
                dateOfEventEnd:
                  type: string
                  format: date-time
                  description: Event end date and time
                  example: "2024-07-15T23:00:00Z"
                status:
                  type: string
                  enum: [CREATED, ACTIVE, COMPLETED, CANCELLED]
                  description: Event status
                  example: "ACTIVE"
                venueLayout:
                  type: object
                  description: Venue layout configuration (JSON object)
                  example: {"sections": ["A", "B", "C"], "capacity": 5000}
                userEmail:
                  type: string
                  format: email
                  description: Email of the user who created the event
                  example: "user@example.com"
                attachmentUrls:
                  type: array
                  description: Array of attachment URLs for AI chatbot knowledge base
                  items:
                    type: string
                    format: uri
                    maxLength: 2048
                  example: ["https://example.com/event-plan.pdf", "https://example.com/venue-map.jpg"]
                attachmentContext:
                  type: string
                  description: Text interpretation/context of attachments for AI chatbot knowledge base
                  maxLength: 10000
                  example: "Event plan includes detailed schedule and safety protocols. Venue map shows emergency exits and first aid stations."
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Event'
                  message:
                    type: string
                    example: "Event updated successfully"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete event
      description: Deletes an event and its associated simulation
      tags:
        - Events
      parameters:
        - name: eventId
          in: path
          required: true
          description: Event ID
          schema:
            type: string
      responses:
        '200':
          description: Event deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Event deleted successfully"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/events/{eventId}/attachments/supported-types:
    get:
      summary: Get supported file types
      description: Returns information about supported file types and processing capabilities
      tags:
        - Events
      parameters:
        - name: eventId
          in: path
          required: true
          description: Event ID (for consistency with API structure)
          schema:
            type: string
            example: "evt_a03bf463-7940-4df0-a612-8e41994784e2"
      responses:
        '200':
          description: Supported file types information
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/SupportedFileTypes'
...
components:
  schemas:
...
    Pagination:
      type: object
      properties:
        currentPage:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 10
        totalItems:
          type: integer
          example: 95
        itemsPerPage:
          type: integer
          example: 10
        hasNextPage:
          type: boolean
          example: true
        hasPreviousPage:
          type: boolean
          example: false
...
  responses:
    BadRequest:
      description: Bad request - validation errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Validation failed"
            code: "VALIDATION_ERROR"
            details:
              - field: "expectedAttendees"
                message: "Expected attendees must be a positive number"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Internal server error"
            code: "INTERNAL_ERROR"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
security:
  - ApiKeyAuth: []
...